+++ test/prompt_decoding_test.dart
@@ -1,6 +1,6 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:ritmus_feedback/src/models/prompt.dart';
-import 'package:ritmus_feedback/src/models/question.dart';
+import 'package:usergist_feedback/src/models/prompt.dart';
+import 'package:usergist_feedback/src/models/question.dart';
 
 void main() {
   test('decodes rating question', () {