+++ test/segment_evaluator_test.dart
@@ -1,6 +1,6 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:ritmus_feedback/src/internal/triggers/segment_evaluator.dart';
-import 'package:ritmus_feedback/src/models/segment.dart';
+import 'package:usergist_feedback/src/internal/triggers/segment_evaluator.dart';
+import 'package:usergist_feedback/src/models/segment.dart';
 
 void main() {
   test('null rules always match', () {